pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aphrodite-loadbalancer"
version = "0.1.0"
description = "Load balancer for Aphrodite API endpoints"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
aphrodite-loadbalancer = "aphrodite_loadbalancer.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]